created app named twubric containing user data like friends, influence, Chirpiness, joining date, score where we can sort user in ascending order based on above data and also remove specific user
